==> R/02_clean-data-raw.R <==
# Clean raw data to unify variable names and formats

library(dplyr)
library(pointblank)

dir_raw <- here::here("data", "raw")
dir_cleaned <- fs::dir_create(here::here("data", "cleaned"))

source(here::here("R", "functions", "extract_design.R"))

# Function: Coalesce registry names ---------------------------------------
# Registry names differ between NDs ICTRP and MSHs ctregistries
# For parity, coalesce to ctregistries
# Note: Disambiguates 3 JPRN registries

disambiguate_registries <- function(reg_df){
  reg_df |>
    rename(registry = source_register) |>
    mutate(registry = case_when(
      registry == "EU Clinical Trials Register" ~ "EudraCT",
      registry == "CRIS" ~ "CRiS",
      registry == "German Clinical Trials Register" ~ "DRKS",
      registry == "REBEC" ~ "ReBec",
      registry == "Netherlands Trial Register" ~ "NTR",

      stringr::str_detect(trn, "JPRN-UMIN") ~ "UMIN-CTR",
      stringr::str_detect(trn, "JPRN-jRCT") ~ "jRCT",
      stringr::str_detect(trn, "JPRN-JapicCTI") ~ "JapicCTI",
      TRUE ~ registry
    ))
}

# In ICTRP, EUCTR TRNs have country codes/text (i.e., tri08064/EUCTR2021-001279-18-Outside/EEA)
# Therefore, often >1 row per EUCTR TRN
# Extract country code to separate variable and remove from TRN
tidy_euctr <- function(reg_df){
  reg_df |>
    mutate(
      euctr_country = if_else(
        stringr::str_detect(trn, "^EUCTR"),
        stringr::str_remove(trn, ".{20}"),
        NA_character_
      ),
      trn = if_else(
        stringr::str_detect(trn, "^EUCTR"),
        stringr::str_extract(trn, "^EUCTR20\\d{2}\\W*0\\d{5}\\W*\\d{2}"),
        trn
      )
    )
}


# Phase 1 Data ------------------------------------------------------------
trials_1 <-
  readr::read_csv(fs::path(dir_raw, "trials_1.csv")) |>
  rename(trn = trialid) |>
  disambiguate_registries() |>
  tidy_euctr()
# registrations_1 <- readr::read_csv(fs::path(dir_raw, "registrations_1.csv"))
# results_1 <- readr::read_csv(fs::path(dir_raw, "results_1.csv"))

readr::write_csv(trials_1, fs::path(dir_cleaned, "trials_1.csv"))


# ICTRP -------------------------------------------------------------------
ictrp_design_sci_title <-
  readr::read_csv(fs::path(dir_raw, "2021-07-01_ictrp_web.csv")) |>
  select(trn = TrialID, study_design = `Study design`, scientific_title = `Scientific title`) |>
  tidy_euctr()

ictrp <-
  readr::read_csv(fs::path(dir_raw, "2021-07-01_ictrp.csv")) |>

  # Clean names
  select(-1, trn = trialid) |>

  disambiguate_registries() |>
  tidy_euctr() |>

  # Add title for EUCTR missing title
  mutate(public_title = if_else(
    trn == "EUCTR2021-002174-52",
    "Essai multicentrique randomisé en ouvert comparant l’efficacité immunologique d’un schéma vaccinal combinant deux vaccins ARNm Covid19 (Pfizer-BioNTech et Moderna) à celle d’une vaccination homologue de chaque vaccin ARNm Covid19 : essai de non infériorité",
    public_title
  )) |>

  # Check that all trials have title
  assertr::assert(assertr::not_na, public_title) |>

  # Check that trials have registration date, except known missing
  # Some REBEC (n=38) and REPEC (n=1) trials are missing registration date
  # REPEC "PER-002-21" seems to no longer exist so no date found
  # REBEC doesn't include completion date so would be excluded anyways
  pointblank::col_vals_not_null(
    vars(date_registration),
    preconditions = ~ . %>% filter(registry != "ReBec" & trn != "PER-002-21")
  ) |>

  # Add in study design and scientific title
  left_join(ictrp_design_sci_title, by = c("trn", "euctr_country")) |>

  # Fill in missing study designs
  # NOTE: ND checked and these are due to his cleaning, and ok
  rows_patch(tribble(
    ~trn, ~study_design,
    "NCT04667780", "Allocation: Randomized. Intervention model: Parallel Assignment. Primary purpose: Treatment. Masking: None (Open Label).",
    "NCT04926428", "Allocation: N/A. Intervention model: Single Group Assignment. Primary purpose: Diagnostic. Masking: None (Open Label).",
    "NCT04933864", "Allocation: Randomized. Intervention model: Parallel Assignment. Primary purpose: Treatment. Masking: Single (Outcomes Assessor).",
  ), by = "trn") |>

  # Extract design
  # NOTE: ReBec not extracted
  extract_design() |>

  # Fill in missing randomization
  # NOTE: Dual coded MSH/ND
  rows_patch(tribble(
    ~trn, ~randomisation,
    "ChiCTR2000029487", "Yes",  # "Third party uses random number table for random grouping"
    "ChiCTR2000030116", "Yes",  # "Generated by a table of random numbers"
    "ChiCTR2000030810", "Yes",  # "Digital random method"
    "ChiCTR2000030773", "No",  #
    "ChiCTR2000030704", "Yes",  # "doctor decides that by envelope method"
    "ChiCTR2000040632", "No",  # "No Randomization was done. It was a single group study"
    "CTRI/2020/10/028486", "Yes",  # "we will randomize the patients into two groups"
    "CTRI/2020/10/028485", "No",  # Method of Generating Random Sequence: Not applicable
    "ISRCTN95933274", "Yes",  # "Randomised controlled trial"
    "ISRCTN34160010", "No",  # "Non randomised study"
    "TCTR20210125002", "No",  # "Single arm"
  ), by = "trn") |>

  # Change to randomization boolean
  mutate(is_randomized = case_when(
    randomisation == "Yes" ~ TRUE,
    randomisation == "No" ~ FALSE,
    TRUE ~ NA
  )) |>
  select(-randomisation)

# There should be dupes only for different EUCTR country codes
# They have different data (i.e., registration dates, titles), so keep for screening checks
janitor::get_dupes(ictrp, trn) |>
  assertr::verify(registry == "EudraCT")

readr::write_csv(ictrp, fs::path(dir_cleaned, "2021-07-01_ictrp.csv"))

# Registries --------------------------------------------------------------

# We use registry data from 2021-07 for phase 3 trial screening. However, NDs scrapes in 2021-07 did not include the entire 2021-07 ICTRP COVID list, so for missing trials we use data from (1) registry data scrapes from 2022-04, if `last_updated` prior to earliest scrape date in 2021-07, and (2) if not available and trial eligible based on ICTRP, we manually check registries for data prior to that date (n = 9, none of which have registrations)
# We also had some cross-registrations missing registry data. We manually gathered historic completion date data in 2022-10 to use to check the inclusion of cross-registered trials across registrations.

# ND scraped various registries between 2021-07-11 and 2021-07-18
REGISTRY_SCRAPE_MIN <- as.Date("2021-07-11")
REGISTRY_SCRAPE_MAX <- as.Date("2021-07-18")

clean_registrations <- function(reg_df){
  reg_df |>
    select(-1, -relevant_comp_date) |>
    rename(trn = trial_id) |>
    mutate(rcd = if_else(!is.na(pcd), pcd, scd), .after = "scd") |>
    mutate(across(
      c(pcd, scd, rcd, last_updated), ~ as.Date(., "%d/%m/%Y")
    )) |>
    mutate(across(c(tabular_results, potential_other_results), ~if_else(. == 1, TRUE, FALSE))) |>
    distinct() |>

    # Add EUCTR prefix
    mutate(trn = if_else(
      stringr::str_detect(trn, "^20\\d{2}\\W*0\\d{5}\\W*\\d{2}"),
      stringr::str_c("EUCTR", trn),
      trn
    )) |>

    # Get registry
    # NOTE: computationally inefficient, rework function
    rowwise() |>
    mutate(registry = ctregistries::which_registry(trn), .after = "trn") |>
    ungroup() |>

    arrange(registry)
}

registries_2107 <-
  readr::read_csv(fs::path(dir_raw, "2021-07_registries.csv")) |>
  clean_registrations() |>
  mutate(registry_scrape_date = "2021-07")

registries_2204 <-
  readr::read_csv(fs::path(dir_raw, "2022-04_registries.csv")) |>

  # Remove invalid trn
  filter(stringr::str_detect(trial_id, "\\?", negate = TRUE)) |>

  clean_registrations() |>
  mutate(registry_scrape_date = "2022-04")

# ND post-hoc collected registry data from 2021-07 for several cross-registrations
registries_2107_manual <-
  readr::read_csv(here::here("data", "manual", "2021-07_registries-manual.csv")) |>
  select(-resolved, -rcd) |>

  # Recalculate rcd in case
  mutate(rcd = if_else(!is.na(pcd), pcd, scd), .after = "scd")

# Confirm that later registry data not in earlier data
semi_join(registries_2107_manual, registries_2107, by = "trn") %>%
  assertr::verify(nrow(.) == 0)

# Confirm that no later registry data with includable registry scrape date not in earlier data
registries_2204 |>
  semi_join(registries_2107_manual, by = "trn") |>
  filter(last_updated < REGISTRY_SCRAPE_MIN) %>%
  assertr::verify(nrow(.) == 0)

# Combine 2021-07 data (auto and manual)
registries_2107_auto_manual <-
  bind_rows(registries_2107, registries_2107_manual)

# Add additional registrations from 2022-04 data last updated prior to 2021-07 scrapes
# May include registrations NOT in 2021-07 ictrp
registries_2107_2204 <-
  registries_2204 |>

  # Limit to registrations updated prior to earliest 2021-07 scrapes
  filter(last_updated < REGISTRY_SCRAPE_MIN) |>

  # Limit to registrations not already in 2021-07 registries
  anti_join(registries_2107_auto_manual, by = "trn") %>%

  # Combine with 2021-07 registries
  bind_rows(registries_2107_auto_manual, .) |>
  arrange(registry)

readr::write_csv(registries_2107_2204, fs::path(dir_cleaned, "2021-07_registries.csv"))

readr::write_csv(registries_2204, fs::path(dir_cleaned, "2022-04_registries.csv"))
